UX: remove global notice to enable encryption
We should defer annoying users till the last possible chance.

This means just as you are composing a message and if we discover you
have encryption setup, and the user you are composing to does as well,
and you are just about to post the message which has lock ticked. At that
point, a modal of "Please enter your encryption password" is reasonable.

Otherwise every 1 time user of the feature is going to be constantly reminded
of a feature they may have only used once.