DEV: Fix test that checks for leaked plaintexts.
During testing, the post is sometimes saved twice, which calls the save
hook twice, generating two different encryption keys and producing two
ciphertexts. This is most probably caused by a bug in mocks, fixture or
the state is not properly reset after previous tests.

The test would fail with an OperationError which was raised because the
post would be decrypted using the wrong key. When saving, only the last
ciphertext would persist, but `putTopicKey` saved the first key, which
caused an incompatibility.

The failure was reproduced only in the 'discourse_test' Docker image.

This commit also updates to the latest Draft API.