FIX: default encryption could be overwritten
The bug was introduced in that commit (message encrypted by default)- https://github.com/discourse/discourse-encrypt/commit/029ef4a63ff1a2b937671b89e0f7b80b00e145f9

Problem was that the user could not overwrite default encryption. So the message was always encrypted when encryption was enabled. We need to allow the user to overwrite that behaviour.