FIX: encrypt works for a user who cannot edit posts (#24)
The bug mentioned here: https://meta.discourse.org/t/min-trust-to-edit-post-setting-interfering-with-discourse-encrypt/163245

Using the guardian was the best and safest approach. However, there is a situation when user cannot edit posts. For example when `min_trust_to_edit_post` is set higher than the user trust level. Therefore we can simply check if the user is author of the post.