FIX: Make sure all user identity promises are eventually resolved (#41)
getDebouncedUserIdentities used to resolve only the first returned
promise. It has been refactored into getDebouncedUserIdentity to
simplify the API and the logic behind it.

The green signature checkmark displayed in the UI is displayed only
after the verification process finishes. This did not happen because
the whole process was stuck waiting for the the promise returned by
getDebouncedUserIdentities to be resolved.