made paper key warning less stark
This warning makes me as a user feel like I've done something wrong. I've changed the warning to make it less stark. Would be nice to be taken to the Generate Paper Key interface first, without the warning, then if the user skips away from it show them the warning.