UX: Use 'paper key' instead of 'passphrase'
This was confusing because the two terms were used interchangeably, but
it was not clear they are synonyms.

Placeholders were added to give user hints about inputs format.

The text in the "Add Another Device" and "Generate Paper Key" modal
used to be the same as because of a bug.