FEATURE: permanently delete a private message
Previously we added a future to set a time for permanent deletion for encrypted messages.

This feature allows user to permanently delete post even if the timer was not set. When the user clicks the delete button, we inform that message will be permanently deleted and create a timer for 1 minute. It allows the user to rollback if clicked by mistake.