FIX: Do more to decrypt encrypted results
Sometimes encrypted results are returned to the user (for example when
searching for "secret", part of the encrypted PM placeholder). Usually,
the encrypted messages cache stored locally contains a decrypted copy.
For these cases, the messages can be decrypted with no additional cost.

This commit also includes various search improvements.