FIX: Fallback to file size to lookup file info
There is no direct way to share information between the the pre-upload
and post-upload stages of the upload pipeline. To workaround this
limitation, the plugin uses a global map indexed by filename, where it
holds file information.

For certain browsers (Safari) and files (with non-ASCII characters in
filename) this method does not work because the filename read by the
browser and the one returned by the server use a different
representation and keys do not match (eg. "fi%CC%88le" vs "f%C3%AFle").
For these files, the plugin will attempt to find the information it
needs by looking up by file size.