FIX: Show activation modal if identity is missing
A user's identity was missing if they had encryption enabled, but the
current device was not activated. To make it easy to use, the user was
prompted to activate the device when they navigated to an encrypted
topic. This behavior changed when `getIdentity` started to return a
rejected promise when user has no identity (it used to return an empty
identity).