FIX: Decrypt titles from generic elements first (#118)
There is a list of CSS selectors that match most of the elements that
contain topic titles. The generic ones can overlap with the more
specific ones. It is best to decrypt the elements matched by generic
ones first and then let the specific ones override them.

This commit fixes the bug that shows two secret icons in header.