FIX: Show activation modal if identity is missing (#117)
User's identity was missing if they had encryption enabled, but the
current device was not activated. To make it easy to use, the user was
prompted to activate the device when they navigated to an encrypted
topic. This behavior changed when `getIdentity` started to return a
rejected promise when user has no identity (it used to return an empty
identity).

This commit removes the underscore prefix from exported function names.
It used to mean that functions could be used only in tests, but now an
error is thrown instead.