DEV: Use new core addComposerUploadPreProcessor API
See https://github.com/discourse/discourse/pull/14222/commits/8261f1e33de76e56abcf5c9d143bb691f9a331d9

Now that we are using Uppy in core, we need a new way to preprocess
files for uploads. This is done with an uppy plugin by convention. This
commit introduces the plugin and uses it along with the new
addComposerUploadPreProcessor API _only_ if enable_experimental_composer_uploader
is true for the site.