DEV: Use new core addComposerUploadPreProcessor API (#131)
See https://github.com/discourse/discourse/pull/14222

Now that we are using Uppy in core, we need a new way to preprocess
files for uploads. This is done with an uppy plugin by convention. This
commit introduces the plugin and uses it along with the new
addComposerUploadPreProcessor API _only_ if enable_experimental_composer_uploader
is true for the site.