FIX: Improve search in encrypted topics (#140)
Filtering has been improved to ignore empty words and order the results
by the number of matched words in query relative to the result of the
length (i.e. closer matches will be at the top of the list).

Encrypted results that were returned by the server will be decrypted and
saved into the cache too.