FIX: Swap fancy_title with decrypted title
The topic title from a topic page used to be replaced with the decrypted
title after it was rendered. This meant that the placeholder title
element was replaced with another element that contained the decrypted
topic title.

This was a problem because it changed the DOM structure and Ember stored
references to elements it rendered, but were replaced. More
specifically, it called `removeChild` to remove an element that did not
exist anymore, which threw an exception.

This solution gets the model object from the topic controller and
replaces its fancy_title with the decrypted title.