FEATURE: Disallow unsafe CSP if plugin is enabled
This plugin adds validation encrypt_enabled and CSP site settings to
make it impossible to enable unsafe CSP settings and the plugin.

Unsafe CSP is a security risk for the plugin because it makes it easier
for potential attackers to inject malicious code that can steal keys or
encrypted conversations.