FIX: Ensure bots can send encrypted messages (#156)
In 1492adb we ensured we created a topic key for all the participants
in the PM. However, since the bots (like the system user) have no way to
generate an encryption key, they can't send encrypted messages anymore.

This bypass the check for bots as we don't really care if they can read
the encrypted message. This allow to programmatically send encrypted
message as the system user for example.