DEV: Dont use saved namespace (#253)
To resolve
```
DEPRECATION: The <discourse@component:plugin-connector::ember19469>#save computed property was just overridden. This removes the computed property and replaces it with a plain value, and has been deprecated
```
we need to not use the saved namespace of `save` as it was overriding a already defined computed function printing the deprecation above. Instead use a more unique `savePreference` action name.