FIX: Don't error when rotating identities of protocol-v0 (#277)
Users who are currently stuck with v0 identities are unable to rotate their identities into v1 because of a bug where the old identity is assumed to be v1 and has signing keys, but v0 identities don't have signing keys. When rotating an identity, the signing keys of the old identity are copied to the new identity to keep old signatures working. However, when the old identity is v0 (which has no signing keys), we attempt to copy the old signature keys (`undefined`) to the new v1 identity, but it ends up with no signing keys and fails to be saved because v1 identities are expected to have signing keys.

This commit fixes the issue by checking for the existence of old signing keys before copying them to the new identity. If there are no signing keys in the old identity, we let the new identity save with the signing keys that are generated with it.