DEV: Modernize topic metadata decryption (#294)
Previously this plugin was using various strategies to watch for DOM updates and then patch in decrypted topic titles and icons. This led to some unfortunate UX quirks (e.g. flickering, double icons, etc.). It was also problematic from a performance/stability point of view, because it introduced an `afterRender` hook on all Ember Components.

This commit extends the work started by 255724ebc5fc3956f26beca09c1f7cb273d76eb2. Decryption is moved from the UI layer to the 'registerTransformation' API which is called by core asynchronously when loading data from the server. This means that, by the time it gets to the rendering, stuff is already decrypted.

For now, this commit just tackles topic titles and the 'encrypted' icon. In future, we will likely want to apply the same strategy to post contents.