FEATURE: Allow users to decrypt existing PMs
Introduces a new site setting: 'allow_decrypting_pms'

When enabled, a new button is added below the title of any encrypted PM. This can be used by any participant to decrypt the entire topic (including uploads) permanently, and turn the topic into a regular unencrypted PM.

Also unskips and fixes-up all the system specs. We no longer run encrypt as part of the Discourse core test suite, so the prior concerns about flakiness should be less of an issue.