FIX: Use database for original upload filenames when decrypting (#349)
The `downloadEncryptedFile` strategy relies on the Content-Disposition header to determine the original filename. That doesn't work on forums with S3 uploads, because cross-origin header access is restricted. This commit switches to looking up the filename from the database instead, so it's always guaranteed to be present. Also bails out if the filename has no extension.